chatbot_backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
flask-socketio==5.3.6
python-socketio==5.11.0

gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23

numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# TEMP DISABLED – was causing build issues
# scikit-learn==1.3.2
